Refactor QuickAppend to write all stream and event changes in a single command.
Quickappend now updates all streams and events as part of a single storageCommand.
The sql is a new function that performs all the stream create/update, version check, and event appends as part of a single sql statement. The only significant way to make this faster on the db side would be to use a binary import to load the data.

The result is appending 1000 events to a new/existing stream is slightly faster, but appending 1 event to 1000 new/existing streams is ~4x faster! Also, the execution time still scales with the number of streams, but its a much flatter scaling.

If you have projections then any performance is completely overshadowed by the impact of writing projections, but this does suggest that a similar bulk upsert for projections will be a good next step